Move interchange format functions to own python file, and fix tests to use module structure.
