+++ primap2/tests/test_data_reading.py
@@ -7,7 +7,7 @@ import pandas as pd
 import pytest
 
 import primap2 as pm2
-import primap2.pm2io._data_reading as pm2io
+import primap2.pm2io as pm2io
 
 from .utils import assert_ds_aligned_equal
 
@@ -24,59 +24,86 @@ DATA_PATH = Path(__file__).parent / "data"
     ],
 )
 def test_convert_ipcc_code_primap_to_primap2(code_in, expected_code_out):
-    assert pm2io.convert_ipcc_code_primap_to_primap2(code_in) == expected_code_out
+    assert (
+        pm2io._data_reading.convert_ipcc_code_primap_to_primap2(code_in)
+        == expected_code_out
+    )
 
 
 def test_convert_ipcc_code_primap_to_primap2_too_short(caplog):
-    assert pm2io.convert_ipcc_code_primap_to_primap2("IPC") == "error_IPC"
+    assert pm2io._data_reading.convert_ipcc_code_primap_to_primap2("IPC") == "error_IPC"
     assert "WARNING" in caplog.text
     assert "Too short to be a PRIMAP IPCC code." in caplog.text
 
 
 def test_convert_ipcc_code_primap_to_primap2_wrong_format(caplog):
-    assert pm2io.convert_ipcc_code_primap_to_primap2("IPD1A") == "error_IPD1A"
+    assert (
+        pm2io._data_reading.convert_ipcc_code_primap_to_primap2("IPD1A")
+        == "error_IPD1A"
+    )
     assert "WARNING" in caplog.text
     assert "Prefix is missing, must be one of 'IPC' or 'CAT'." in caplog.text
 
 
 def test_convert_ipcc_code_primap_to_primap2_first_lvl(caplog):
-    assert pm2io.convert_ipcc_code_primap_to_primap2("IPCA1") == "error_IPCA1"
+    assert (
+        pm2io._data_reading.convert_ipcc_code_primap_to_primap2("IPCA1")
+        == "error_IPCA1"
+    )
     assert "WARNING" in caplog.text
     assert "No digit found on first level." in caplog.text
 
 
 def test_convert_ipcc_code_primap_to_primap2_second_lvl(caplog):
-    assert pm2io.convert_ipcc_code_primap_to_primap2("IPC123") == "error_IPC123"
+    assert (
+        pm2io._data_reading.convert_ipcc_code_primap_to_primap2("IPC123")
+        == "error_IPC123"
+    )
     assert "WARNING" in caplog.text
     assert "No letter found on second level." in caplog.text
 
 
 def test_convert_ipcc_code_primap_to_primap2_third_lvl(caplog):
-    assert pm2io.convert_ipcc_code_primap_to_primap2("IPC1AC") == "error_IPC1AC"
+    assert (
+        pm2io._data_reading.convert_ipcc_code_primap_to_primap2("IPC1AC")
+        == "error_IPC1AC"
+    )
     assert "WARNING" in caplog.text
     assert "No number found on third level." in caplog.text
 
 
 def test_convert_ipcc_code_primap_to_primap2_fourth_lvl(caplog):
-    assert pm2io.convert_ipcc_code_primap_to_primap2("IPC1A2_") == "error_IPC1A2_"
+    assert (
+        pm2io._data_reading.convert_ipcc_code_primap_to_primap2("IPC1A2_")
+        == "error_IPC1A2_"
+    )
     assert "WARNING" in caplog.text
     assert "No letter found on fourth level." in caplog.text
 
 
 def test_convert_ipcc_code_primap_to_primap2_fifth_lvl(caplog):
-    assert pm2io.convert_ipcc_code_primap_to_primap2("IPC1A2BB") == "error_IPC1A2BB"
+    assert (
+        pm2io._data_reading.convert_ipcc_code_primap_to_primap2("IPC1A2BB")
+        == "error_IPC1A2BB"
+    )
     assert "WARNING" in caplog.text
     assert "No digit found on fifth level." in caplog.text
 
 
 def test_convert_ipcc_code_primap_to_primap2_sixth_lvl(caplog):
-    assert pm2io.convert_ipcc_code_primap_to_primap2("IPC1A2B3X") == "error_IPC1A2B3X"
+    assert (
+        pm2io._data_reading.convert_ipcc_code_primap_to_primap2("IPC1A2B3X")
+        == "error_IPC1A2B3X"
+    )
     assert "WARNING" in caplog.text
     assert "No number found on sixth level." in caplog.text
 
 
 def test_convert_ipcc_code_primap_to_primap2_after_sixth_lvl(caplog):
-    assert pm2io.convert_ipcc_code_primap_to_primap2("IPC1A2B33A") == "error_IPC1A2B33A"
+    assert (
+        pm2io._data_reading.convert_ipcc_code_primap_to_primap2("IPC1A2B33A")
+        == "error_IPC1A2B33A"
+    )
     assert "WARNING" in caplog.text
     assert "Chars left after sixth level." in caplog.text
 
@@ -91,25 +118,35 @@ def test_convert_ipcc_code_primap_to_primap2_after_sixth_lvl(caplog):
     ],
 )
 def test_convert_unit_primap_to_primap2(unit_in, entity_in, expected_unit_out):
-    assert pm2io.convert_unit_primap_to_primap2(unit_in, entity_in) == expected_unit_out
+    assert (
+        pm2io._data_reading.convert_unit_primap_to_primap2(unit_in, entity_in)
+        == expected_unit_out
+    )
 
 
 def test_convert_unit_primap_to_primap2_no_prefix(caplog):
     assert (
-        pm2io.convert_unit_primap_to_primap2("CO2eq", "FGASES") == "error_CO2eq_FGASES"
+        pm2io._data_reading.convert_unit_primap_to_primap2("CO2eq", "FGASES")
+        == "error_CO2eq_FGASES"
     )
     assert "WARNING" in caplog.text
     assert "No unit prefix matched for unit." in caplog.text
 
 
 def test_convert_unit_primap_to_primap2_unit_empty(caplog):
-    assert pm2io.convert_unit_primap_to_primap2("", "FGASES") == "error__FGASES"
+    assert (
+        pm2io._data_reading.convert_unit_primap_to_primap2("", "FGASES")
+        == "error__FGASES"
+    )
     assert "WARNING" in caplog.text
     assert "Input unit is empty. Nothing converted." in caplog.text
 
 
 def test_convert_unit_primap_to_primap2_entity_empty(caplog):
-    assert pm2io.convert_unit_primap_to_primap2("GgCO2eq", "") == "error_GgCO2eq_"
+    assert (
+        pm2io._data_reading.convert_unit_primap_to_primap2("GgCO2eq", "")
+        == "error_GgCO2eq_"
+    )
     assert "WARNING" in caplog.text
     assert "Input entity is empty. Nothing converted." in caplog.text
 
@@ -139,7 +176,9 @@ def test_convert_unit_primap_to_primap2_entity_empty(caplog):
     ],
 )
 def test_metadata_for_variable(unit, entity, expected_attrs):
-    assert pm2io.metadata_for_variable(unit, entity) == expected_attrs
+    assert (
+        pm2io._interchange_format.metadata_for_variable(unit, entity) == expected_attrs
+    )
 
 
 @pytest.mark.parametrize(
@@ -151,7 +190,7 @@ def test_metadata_for_variable(unit, entity, expected_attrs):
     ],
 )
 def test_convert_entity_gwp_primap(entity_pm1, entity_pm2):
-    assert pm2io.convert_entity_gwp_primap(entity_pm1) == entity_pm2
+    assert pm2io._data_reading.convert_entity_gwp_primap(entity_pm1) == entity_pm2
 
 
 def test_read_wide_csv_file(tmp_path):
@@ -676,7 +715,7 @@ def test_convert_dataframe_units_primap_to_primap2():
     file_expected = DATA_PATH / "test_convert_dataframe_units_primap_to_primap2.csv"
     df_expected = pd.read_csv(file_expected, index_col=0)
     df = pd.read_csv(file_input)
-    df_converted = pm2io.convert_dataframe_units_primap_to_primap2(
+    df_converted = pm2io._data_reading.convert_dataframe_units_primap_to_primap2(
         df, unit_col="unit", entity_col="gas"
     )
     pd.testing.assert_frame_equal(df_converted, df_expected)
